make sure resolvconf path exist
Systemd will fail to set up mount namespacing if a path in
InaccessiblePaths does not exist. By using '-' as a prefix for the path,
Systemd will ignore it if the path does not exist. This means the path
is not set as inaccessible, so the path *will* be accessible to the
service if it is created after the service is started. Creating an empty
resolvconf directory beforehand allows us to make the path inaccessible
in case the resolvconf service should somehow activate while our
confined service is still running. Deleting the inaccessible resolvconf
directory and making a new one while the confined service is running
*will* make the newly created directory accessible to the confined service.
Though untested, it is assumed that the resolvconf service would not
replace the directory by deleting it, in which case the path would
remain inaccessible as desired.